refactor: move app code to dedicated directory
